Refactor servers and use webpack middleware
