+++ bin/server.js
@@ -4,11 +4,15 @@ require('../server.babel');
 
 const config = require('../config').default;
 const appName = config.get('currentApp');
+
 const server = require(`../src/${appName}/server`).default;
 
-const port = config.get('serverPort');
-const host = config.get('serverHost');
 const env = config.get('env');
+
+const port = env === 'production' ?
+  config.get('serverPort') : config.get('devServerPort');
+const host = env === 'production' ?
+  config.get('serverHost') : config.get('devServerHost');
 
 server.listen(port, host);
