+++ config/index.js
@@ -2,6 +2,7 @@ import path from 'path';
 
 const config = new Map();
 
+// Default to production unless overridden.
 config.set('env', process.env.NODE_ENV || 'production');
 
 config.set('globals', {
@@ -11,10 +12,11 @@ config.set('globals', {
 
 config.set('basePath', path.resolve(__dirname, '../'));
 
-config.set('serverHost', process.env.NODE_HOST || '127.0.0.1');
-config.set('serverPort', process.env.NODE_PORT || 4000);
-config.set('webpackHost', process.env.NODE_HOST || '127.0.0.1');
-config.set('webpackPort', 3000);
+config.set('serverHost', process.env.SERVER_HOST || '127.0.0.1');
+config.set('serverPort', process.env.SERVER_PORT || 4000);
+
+config.set('devServerHost', process.env.SERVER_HOST || '127.0.0.1');
+config.set('devServerPort', process.env.SERVER_PORT || 3000);
 
 const APP_NAMES = ['search', 'disco'];
 if (APP_NAMES.indexOf(process.env.APP_NAME) < 0) {
