+++ package.json
@@ -5,15 +5,15 @@
   "main": "index.js",
   "private": true,
   "scripts": {
-    "start": "npm run version-check && node bin/server.js",
-    "dev": "npm run start & webpack-dev-server --progress --color --port 3000",
+    "start": "npm run version-check && NODE_PATH=$NODE_PATH:. node bin/server.js",
+    "dev": "NODE_ENV=development npm run start",
     "dev:search": "APP_NAME=search npm run dev",
     "dev:disco": "APP_NAME=disco npm run dev",
     "test": "npm run version-check && npm run lint && npm run jscs",
     "lint": "eslint .",
     "eslint": "npm run lint",
     "jscs": "jscs .",
-    "build": "npm run version-check && webpack --progress --color -p --config webpack.prod.config.js",
+    "build": "npm run version-check && webpack --progress --color -p --config webpack.config.js",
     "version-check": "node bin/version-check.js"
   },
   "repository": {
@@ -51,7 +51,9 @@
     "semver": "5.1.0",
     "shelljs": "0.6.0",
     "webpack": "1.12.14",
+    "webpack-dev-middleware": "1.5.1",
     "webpack-dev-server": "1.14.1",
+    "webpack-hot-middleware": "2.7.1",
     "webpack-isomorphic-tools": "2.2.30"
   }
 }
