--- src/core/server.js
@@ -1,7 +1,0 @@
-import baseServer from './baseServer';
-
-import routes from './routes';
-
-const app = baseServer(routes);
-export default app;
-
