+++ src/core/server/index.js
@@ -1,0 +1,7 @@
+import baseServer from './base';
+
+import routes from '../routes';
+
+const app = baseServer(routes);
+
+export default app;
