+++ src/search/server.js
@@ -1,7 +1,7 @@
-import baseServer from '../core/baseServer';
+import baseServer from '../core/server/base';
 
 import routes from './routes';
 
 const app = baseServer(routes);
-export default app;
 
+export default app;
