--- webpack.prod.config.js
@@ -1,4 +1,0 @@
-// This allows the webpack conf to use ES6.
-require('./server.babel');
-
-module.exports = exports = require('./config/webpack.prod.config');
